Crate tree_dp

Re-rooted tree DP.

NOTE: The parent will automatically be removed.(already removed is also OK)

Notation

$G = (V, E; \ r)$ is a tree.

$G _ x$ is the set of children of $x$.

$P _ x$ is the parent of $x$ (undefined if $x = r$).

Math	Rust	Description

$r$	root	The root of the tree

$X$	Op::Value	DP value

$A$	Op::Acc, Op::identity(), Op::mul()	the monoid of converted values

$W$	Op::VertexWeight	vertex weight (usually be ())

$f: X \to A$	Op::f()	Conversion

$g ^ w: A \to X \ (w \in W)$	Op::g()	Conversion

Usual tree DP

$\mathrm{dp}$ is the recursively defined values of the subtree rooted at $x$. $\mathrm{ep}$ is the converted values of them.

$$ \begin{aligned} \mathrm{dp} _ x &= g ^ { W _ x } \left( \prod _ { y \in G _ x } \mathrm{ep} _ y \right) \in X \\ \mathrm{ep} _ x &= f \left(\mathrm{dp} _ x\right) \in A \end{aligned} $$

Re-rooted tree DP

Let $\mathrm{dp}’(x)$ be the result on the tree rooted at $x$, and with the strict subtree of $x$ removed.

$\mathrm{fp}$ is the converted values of the parent (if none, $1 \in A$) ($y \in G _ x$):

$$ \mathrm{fp}(y) = \begin{cases} 1 & \text{if $y = r$} \\ f\left(\mathrm{dp}’(x)\right) \in A & \text{otherwise} \end{cases} \in A $$

This is computed according to the following recursive formula ($y \in G _ x$):

$$ \mathrm{fp}(y) = \begin{cases} 1 & \text{if $y = r$} \\ \left ( f \circ g ^ { W _ x } \right ) \left ( \mathrm{fp}(x) \cdot \prod _ { y’ \in G _ x \setminus \{y\} } \mathrm{ep}(y’) \right ) & \text{otherwise} \end{cases} \in A $$

$\mathrm{gp}$ is the value of the total tree re-rooted at $x$:

$$ \mathrm{gp}(x) = g ^ { W _ x } \left(\prod_{y \in G _ x \cup P _ x} \mathrm{ep}(y) \right) \in X $$

This is computed according to the following recursive formula:

$$ \mathrm{gp}(x) = g ^ { W _ x } \left( \mathrm{fp}(x) \cdot \left( \prod _ { y \in G _ x } \mathrm{ep}(y) \right) \right) \in X $$

Traits

Op

An operation for re-rooted tree DP.

Functions

rerooted_tree_dp

Given a tree $G = (V, E; \ r)$ and vertex weights $W$, returns the four DP arrays: $\mathrm{dp}$, $\mathrm{ep}$, $\mathrm{fp}$, and $\mathrm{gp}$.